// anmeldetool-web/src/main/web/apps/at-app/src/app/core/model/IAnlassLink.ts
import { MeldeStatusEnum } from './MeldeStatusEnum';

export type IAnlassLink = {
  kategorie: string;
  meldeStatus?: MeldeStatusEnum;
  anlassId: string;
  teilnehmerId: string;
  organisationId: string;
  dirty?: boolean;
  startnummer?: number;
  abteilung?: string;
  startgeraet?: string;
  anlage?: string;
};