// anmeldetool-web/src/main/web/apps/at-app/src/app/core/model/IAnlassSummary.ts
export type IAnlassSummary = {
  anlassId: string;
  organisationsId: string;
  startet: boolean;
  verlaengerungsDate: Date;
  startendeBr1: number;
  startendeK1: number;
  startendeK2: number;
  startendeK3: number;
  startendeK4: number;
  startendeK5: number;
  startendeK5A: number;
  startendeK5B: number;
  startendeK6: number;
  startendeK7: number;
  startendeKD: number;
  startendeKH: number;
  startendeBr2: number;
  gemeldeteBr1: number;
  gemeldeteBr2: number;
  br1Ok: boolean;
  br2Ok: boolean;
};